dltexport: Cancel message export implemented             Add commandline progress output dltindexer: local variable covered parameter              more detailed debug output              fix progressbar overflow mainwindow: avoid error output ERROR: bytesRcvd <= 0             when ECU in dlt file is given             dlt injection: fix error when UTF8 ( chinese character)             transmitted settingsdialog: avoid crash when resetting settings due to version change version.h: set version to 2.20.1 <Signed-off-by: Gernot Wirschal <gernot.wirschal@bmw.de>

+++ src/dltfileindexer.h
@@ -185,10 +185,10 @@ private:
 signals:
 
     // the maximum progress value
-    void progressMax(quint64 index);
+    void progressMax(int index);
 
     // the current progress value
-    void progress(quint64 index);
+    void progress(int index);
 
     // progress text change fro different parts
     void progressText(QString text);
